fitting refactor
Deleted src/brutus/fitting.py and coverage.xml as part of codebase cleanup (the latter is now ignored). Updated .gitignore to exclude coverage artifacts. Made changes to analysis modules and related tests to reflect the removal of fitting.py, which is now fully refactored elsewhere in the code.

Also, tried to improve the runtime for the fitting refactor to use more numba, and also fixed a few bugs involved with the covariance matrices (of course).